+++ src/main/java/kdt/prgms/springbootboard/global/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package kdt.prgms.springbootboard.global.config.error;
+package kdt.prgms.springbootboard.global.error;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
